+++ CHANGELOG.md
@@ -11,7 +11,11 @@ Anticipated release: July 9, 2021
 
 #### ⚙️ Behind the scenes
 
+
 - install cypress for end-to-end testing ([#3226])
+
+- Created endpoint to allow a user to switch states ([#3030])
+
 
 # Previous releases
 
@@ -19,3 +23,4 @@ See our [release history](https://github.com/CMSgov/eAPD/releases)
 
 [#3226]: https://github.com/CMSgov/eAPD/issues/3226
 [#2996]: https://github.com/CMSgov/eAPD/issues/2996
+[#3030]: https://github.com/CMSgov/eAPD/issues/3030